Use floats instead of double, faster to run Starting using value trees to read parameters, let's see if this works out properly!

+++ Dynamics/ATKColoredCompressor/Source/PluginProcessor.h
@@ -72,12 +72,12 @@ private:
     
   ATK::InPointerFilter<float> inFilter;
   ATK::OutPointerFilter<float> outFilter;
-  ATK::PowerFilter<double> powerFilter;
-  ATK::AttackReleaseFilter<double> attackReleaseFilter;
-  ATK::GainFilter<ATK::GainColoredCompressorFilter<double>> gainCompressorFilter;
-  ATK::ApplyGainFilter<double> applyGainFilter;
-  ATK::VolumeFilter<double> volumeFilter;
-  ATK::DryWetFilter<double> drywetFilter;
+  ATK::PowerFilter<float> powerFilter;
+  ATK::AttackReleaseFilter<float> attackReleaseFilter;
+  ATK::GainFilter<ATK::GainColoredCompressorFilter<float>> gainCompressorFilter;
+  ATK::ApplyGainFilter<float> applyGainFilter;
+  ATK::VolumeFilter<float> volumeFilter;
+  ATK::DryWetFilter<float> drywetFilter;
 
   AudioProcessorValueTreeState parameters;
   int sampleRate;
